interpolation/requirements.txt
quixstreams==2.8.1
python-dotenv
pandas